requirements.txt
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
dj-rest-auth==2.1.3
Django==3.1.5
django-filebrowser-no-grappelli==3.8.0
django-tinymce4-lite==1.8.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
jsmin==2.2.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
